+++ app/src/main/java/org/mozilla/fenix/BrowserDirection.kt
@@ -16,7 +16,6 @@ import androidx.annotation.IdRes
 enum class BrowserDirection(@IdRes val fragmentId: Int) {
     FromGlobal(0),
     FromHome(R.id.homeFragment),
-    FromSearch(R.id.searchFragment),
     FromSearchDialog(R.id.searchDialogFragment),
     FromSettings(R.id.settingsFragment),
     FromSyncedTabs(R.id.syncedTabsFragment),
