+++ app/src/main/res/navigation/nav_graph.xml
@@ -24,10 +24,6 @@
         app:popUpToInclusive="false" />
 
     <action
-        android:id="@+id/action_global_search"
-        app:destination="@id/searchFragment" />
-
-    <action
         android:id="@+id/action_global_search_dialog"
         app:destination="@id/searchDialogFragment" />
 
@@ -145,7 +141,7 @@
 
     <dialog
         android:id="@+id/searchDialogFragment"
-        android:name="org.mozilla.fenix.searchdialog.SearchDialogFragment"
+        android:name="org.mozilla.fenix.search.SearchDialogFragment"
         tools:layout="@layout/fragment_search_dialog">
         <argument
             android:name="session_id"
@@ -161,25 +157,6 @@
             android:defaultValue="NONE"
             app:argType="org.mozilla.fenix.components.metrics.Event$PerformedSearch$SearchAccessPoint" />
     </dialog>
-
-    <fragment
-        android:id="@+id/searchFragment"
-        android:name="org.mozilla.fenix.search.SearchFragment"
-        tools:layout="@layout/fragment_search">
-        <argument
-            android:name="session_id"
-            app:argType="string"
-            app:nullable="true" />
-        <argument
-            android:name="pastedText"
-            android:defaultValue="@null"
-            app:argType="string"
-            app:nullable="true" />
-        <argument
-            android:name="search_access_point"
-            android:defaultValue="NONE"
-            app:argType="org.mozilla.fenix.components.metrics.Event$PerformedSearch$SearchAccessPoint" />
-    </fragment>
 
     <fragment
         android:id="@+id/recentlyClosedFragment"
@@ -200,11 +177,6 @@
         android:name="org.mozilla.fenix.browser.BrowserFragment"
         app:exitAnim="@anim/fade_out"
         tools:layout="@layout/fragment_browser">
-        <action
-            android:id="@+id/action_browserFragment_to_searchFragment"
-            app:destination="@id/searchFragment"
-            app:enterAnim="@anim/fade_in_up"
-            app:popExitAnim="@anim/fade_out_down" />
         <argument
             android:name="activeSessionId"
             app:argType="string"
